+++ tina-rpc-tokio/tests/bridge.rs
@@ -63,6 +63,9 @@ impl<T> Mailbox<T> for EMailbox<T> {
     fn recv(&self) -> Option<T> {
         self.queue.borrow_mut().pop_front()
     }
+    fn is_empty(&self) -> bool {
+        self.queue.borrow().is_empty()
+    }
     fn close(&self) {
         self.closed.set(true);
     }