+++ tina-runtime/src/tests.rs
@@ -188,6 +188,10 @@ impl<T> Mailbox<T> for TestMailbox<T> {
         self.queue.borrow_mut().pop_front()
     }
 
+    fn is_empty(&self) -> bool {
+        self.queue.borrow().is_empty()
+    }
+
     fn close(&self) {
         self.closed.set(true);
     }